Update for the latest released version, make it more readable and easier to use
