+++ .gitignore
@@ -1,4 +1,3 @@
 __pycache__/
 *.pyc
 build/
-paretoflow.egg-info/
