+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include long_description.md
+include README.md
