+++ paretoflow.egg-info/requires.txt
@@ -1,0 +1,5 @@
+pymoo>=0.6.0
+numpy>=1.23.2
+scipy>=1.10.1
+torch>=2.0.1
+tqdm>=4.66.5
